requirements.txt
python-telegram-bot[webhooks]==20.3
emoji==2.8.0
